Move simulation parameters down to the function level Fixes #101
