+++ policyengine/__init__.py
@@ -1,1 +1,1 @@
-from .simulation import Simulation, SimulationOptions
+from .policyengine import PolicyEngine
