+++ policyengine/outputs/household/comparison/calculate_household_comparison.py
@@ -2,7 +2,7 @@
 
 import typing
 
-from policyengine import Simulation
+from policyengine import PolicyEngine
 
 from pydantic import BaseModel
 from policyengine.utils.calculations import get_change
@@ -27,7 +27,7 @@ class HouseholdComparison(BaseModel):
 
 
 def calculate_household_comparison(
-    simulation: Simulation,
+    engine: PolicyEngine,
 ) -> HouseholdComparison:
     """Calculate comparison statistics between two household scenarios."""
     if not simulation.is_comparison:
