+++ policyengine/outputs/household/single/calculate_single_household.py
@@ -2,7 +2,7 @@
 
 import typing
 
-from policyengine import Simulation
+from policyengine import PolicyEngine
 
 from pydantic import BaseModel
 from policyengine_core.simulations import Simulation as CountrySimulation
@@ -33,7 +33,7 @@ class SingleHousehold(BaseModel):
 
 
 def calculate_single_household(
-    simulation: Simulation,
+    engine: PolicyEngine,
 ) -> SingleHousehold:
     """Calculate household statistics for a single household scenario."""
     if simulation.is_comparison:
