+++ policyengine/outputs/macro/comparison/charts/budget_by_program.py
@@ -2,14 +2,14 @@ import plotly.express as px
 import plotly.graph_objects as go
 import typing
 
-from policyengine import Simulation
+from policyengine import PolicyEngine
 
 from pydantic import BaseModel
 from policyengine.utils.charts import *
 
 
 def create_budget_program_comparison_chart(
-    simulation: Simulation,
+    engine: PolicyEngine,
 ) -> go.Figure:
     """Create a budget comparison chart."""
     if not simulation.is_comparison:
