+++ policyengine/outputs/macro/comparison/charts/inequality.py
@@ -2,14 +2,14 @@ import plotly.express as px
 import plotly.graph_objects as go
 import typing
 
-from policyengine import Simulation
+from policyengine import PolicyEngine
 
 from pydantic import BaseModel
 from policyengine.utils.charts import *
 
 
 def create_inequality_chart(
-    simulation: Simulation,
+    engine: PolicyEngine,
     relative: bool,
 ) -> go.Figure:
     """Create a budget comparison chart."""
